[SC-72] refactor app for heroku deployment
