--- backend/package.json
@@ -1,19 +1,0 @@
-{
-  "name": "social-connect-backend",
-  "version": "1.0.0",
-  "description": "social-connect-app backend",
-  "main": "app.js",
-  "scripts": {
-    "start": "node app.js"
-  },
-  "author": "social-connect-team",
-  "license": "MIT",
-  "dependencies": {
-    "express": "4.17.1",
-    "socket.io": "2.3.0"
-  },
-  "engines": {
-    "npm": "6.x",
-    "node": "12.x"
-  }
-}
