--- frontend/package.json
@@ -1,41 +1,0 @@
-{
-  "name": "social-connect-frontend",
-  "version": "1.0.0",
-  "description": "social-connect-app frontend",
-  "main": "/src/index.js",
-  "scripts": {
-    "start": "react-scripts start"
-  },
-  "author": "social-connect-team",
-  "license": "MIT",
-  "dependencies": {
-    "@material-ui/core": "^4.11.0",
-    "@material-ui/icons": "^4.9.1",
-    "@testing-library/jest-dom": "^4.2.4",
-    "@testing-library/react": "^9.5.0",
-    "@testing-library/user-event": "^7.2.1",
-    "chat-ui-react": "^0.2.10",
-    "chess.js": "^0.11.0",
-    "chessboardjsx": "^2.4.2",
-    "react": "^16.13.1",
-    "react-dom": "^16.13.1",
-    "react-router-dom": "^5.2.0",
-    "react-scripts": "3.4.3",
-    "socket.io-client": "^2.3.1"
-  },
-  "engines": {
-    "npm": "6.x"
-  },
-  "browserslist": {
-    "production": [
-      ">0.2%",
-      "not dead",
-      "not op_mini all"
-    ],
-    "development": [
-      "last 1 chrome version",
-      "last 1 firefox version",
-      "last 1 safari version"
-    ]
-  }
-}
