+++ src/socket.js
@@ -1,0 +1,5 @@
+import io from "socket.io-client";
+
+var socket = io();
+console.log(socket);
+export { socket };
