+++ src/styles/HostViewStyles.js
