+++ src/styles/PlayPageStyles.js
